# Conflicts:
#	client/components/pages/configuration/users/CreateUserDialog.vue
#	client/utils/forms.ts
#	client/utils/websocketEvents.ts

+++ server/src/main/kotlin/com/kamelia/hedera/rest/auth/SessionManager.kt
@@ -179,7 +179,7 @@ data class UserState(
         role,
         enabled,
         currentDiskQuota,
-        if(maximumDiskQuota > 0.toLong()) currentDiskQuota.toDouble() / maximumDiskQuota.toDouble() else 0.toDouble(),
+        if(maximumDiskQuota > 0L) currentDiskQuota.toDouble() / maximumDiskQuota.toDouble() else 0.0,
         maximumDiskQuota,
         createdAt.toString(),
     )
